Modify the Module Name, Create the Lib Directory with packages decrypt, encrypt and commonFunc
Modify the module name to github.com/jaimejcs/go_esteganografia.
Add the packages encrypt, decrypt and commonFunc to the lib directory to module the project.
Add a file to test if the commands are working correctly.
Modify the name of var commands in cmd directory to avoid conflicts with package imports.